move crop logic to tool.py
